requirements.txt
asgiref==3.5.2
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
future==0.18.2
Markdown==3.3.7
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
pytz==2022.1
serializers==0.2.4
sqlparse==0.4.2
toml==0.10.2
